component cardProduct refactor
